% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters-auto.R
\name{ch01}
\alias{ch01}
\alias{ch1}
\title{Print examples of chapter 1 of 'R for Dummies'.}
\usage{
ch01()

ch1()
}
\description{
To print a listing of all examples of a chapter, use \code{ch1()}. 
To run all the examples of \code{ch1()}, use \code{example(ch1)}.
}
\examples{
# Chapter 1 - Introducing R: The Big Picture

# Recognizing the Benefits of Using R

## It comes as free, open-source code

### It runs anywhere

### It supports extensions

### It provides an engaged community

### It connects with other languages


# Looking At Some of the Unique Features of R

## Performing multiple calculations with vectors

x <- 1:5
x
x + 2
x + 6:10

## Processing more than just statistics

## Running code without a compiler




}
\seealso{
\code{\link{toc}}

Other Chapters: \code{\link{ch02}}, \code{\link{ch03}},
  \code{\link{ch04}}, \code{\link{ch05}},
  \code{\link{ch06}}, \code{\link{ch07}},
  \code{\link{ch08}}, \code{\link{ch09}},
  \code{\link{ch10}}, \code{\link{ch11}},
  \code{\link{ch12}}, \code{\link{ch13}},
  \code{\link{ch14}}, \code{\link{ch15}},
  \code{\link{ch16}}, \code{\link{ch17}},
  \code{\link{ch18}}, \code{\link{ch19}},
  \code{\link{ch20}}
}

