% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generateChapters}
\alias{generateChapters}
\title{Generate code and examples for all chapters.}
\usage{
generateChapters(chapters = 1:20, path, file = "R/chapters-auto.R")
}
\arguments{
\item{chapters}{Chapter numbers}

\item{path}{Path to package}

\item{file}{Name of file containing ch() functions}
}
\description{
Generate code and examples for all chapters.
}
\keyword{Internal}

