% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readExampleFile}
\alias{readExampleFile}
\title{Read cleaned chapter code from /inst/scripts/2-clean}
\usage{
readExampleFile(ch)
}
\arguments{
\item{ch}{Chapter number}
}
\description{
Read cleaned chapter code from /inst/scripts/2-clean
}
\keyword{Internal}

