% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfordummies-package.r
\docType{package}
\name{rfordummies}
\alias{rfordummies}
\alias{rfordummies-package}
\title{A package to accompany the book "R for Dummies".}
\description{
This package contains all the code examples in the book 'R for Dummies' (2nd edition)
}
\references{
de Vries, A. , & Meys, J. (2012). <i>R for dummies</i>. Chichester: Wiley. \url{https://rfordummies.com/}, ISBN-13: 978-1119962847.
}
