% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveElements.R
\name{saveElements}
\alias{saveElements}
\title{Saves a copy of the periodic table of elements as excel or csv file.}
\usage{
saveElements(outfile, type = c("excel", "csv"))
}
\arguments{
\item{outfile}{File name}

\item{type}{Either excel or csv}
}
\description{
Saves a copy of the periodic table of elements as excel or csv file.
}
\examples{
if(require("XLConnect")){
    saveElements(file.path(tempdir(), "elements.xlsx"))
    saveElements(file.path(tempdir(), "elements.csv"), type = "csv")
    list.files(tempdir(), pattern = "xlsx|csv", full.names = TRUE)
}

}

