% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ch12}
\alias{ch12}
\title{Print examples of chapter 12 of 'R for Dummies'.}
\usage{
ch12()
}
\description{
To print a listing of all examples of a chapter, use \code{ch12()}. To run all the examples of \code{ch12()}, use \code{example(ch12)}.
}
\examples{
# Chapter 12
# Getting Data into and out of R

# NOTE : Most of the code depends on actions, directories
# and the presence of files. Code that isn't runnable is
# commented out.

# Getting Data into R

## Entering data in the R text editor

elements <- data.frame()
# elements <- edit(elements)

# print(elements)

## Using the Clipboard to copy and paste
# Reminder : This only works on Windows

## # x <- readClipboard()
# x
## # x <- readClipboard()
# x
# x <- read.table(file = "clipboard", sep = "\\t", header=TRUE)
# x

## Reading data in CSV files

### Using read.csv() to import data

# elements <- read.csv(file.path("f:", "elements.csv"))
# str(elements)
# elements <- read.csv(file.path("f:", "elements.csv"), stringsAsFactors=FALSE)
# str(elements)

### Using read.table() to import tabular data in text files

## Reading data from Excel
## # install.packages("XLConnect")
# library("XLConnect")
# excel.file <- file.path("~/Elements.xlsx")

# elements <- readWorksheetFromFile(excel.file, sheet=1)
# elements <- readWorksheetFromFile(excel.file, sheet="Sheet1")

## Working with other data types

# library(foreign)
# read.spss(file="location/of/myfile")

# Getting Your Data out of R


# writeClipboard(names(iris))

# write.table(head(iris), file="clipboard", sep="\\t", row.names=FALSE)

# Working with Files and Folders

## Understanding the working directory
getwd()

# setwd("F:/git/roxygen2")
# getwd()
# setwd("F:\\\\git\\\\stringr")
# getwd()

file.path("f:", "git", "surveyor")

# setwd(file.path("F:", "git", "roxygen2"))
# getwd()

file.path("F:", "git", "roxygen2", "roxygen2", "README.md" )

## Manipulating files

# list.files(file.path("F:", "git", "roxygen2"))
my.file <- tempfile()
my.file
write.csv(iris, file=my.file)
list.files(tempdir())

file.iris <- read.csv(my.file)

file.remove(my.file)
list.files(tempdir())


}
\seealso{
toc

Other Chapters: \code{\link{ch01}}, \code{\link{ch1}};
  \code{\link{ch02}}, \code{\link{ch2}};
  \code{\link{ch03}}, \code{\link{ch3}};
  \code{\link{ch04}}, \code{\link{ch4}};
  \code{\link{ch05}}, \code{\link{ch5}};
  \code{\link{ch06}}, \code{\link{ch6}};
  \code{\link{ch07}}, \code{\link{ch7}};
  \code{\link{ch08}}, \code{\link{ch8}};
  \code{\link{ch09}}, \code{\link{ch9}};
  \code{\link{ch10}}; \code{\link{ch11}};
  \code{\link{ch13}}; \code{\link{ch14}};
  \code{\link{ch15}}; \code{\link{ch16}};
  \code{\link{ch17}}; \code{\link{ch18}};
  \code{\link{ch19}}; \code{\link{ch20}}
}

