% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters-auto.R
\name{ch20}
\alias{ch20}
\title{Print examples of chapter 20 of 'R for Dummies'.}
\usage{
ch20()
}
\description{
To print a listing of all examples of a chapter, use \code{ch20()}.
To run all the examples of \code{ch20()}, use \code{example(ch20)}.
}
\examples{
if (interactive()) {
# Chapter 20 - Ten Tips on Working with Packages

## Poking Around the Nooks and Crannies of CRAN

options("repos" = c(CRAN = "https://cran.ma.imperial.ac.uk/"))

## Finding Interesting Packages

## Installing Packages

\dontrun{
install.packages("fortunes")
}

## Loading Packages

library("fortunes")

## Reading the Package Manual and Vignette

library(help=fortunes)
\dontrun{
vignette("fortunes")
}

## Updating Packages

\dontrun{
update.packages()
}

## Unloading Packages

search()
detach(package:fortunes, unload=TRUE)

## Forging Ahead with R-Forge

\dontrun{
install.packages("data.table", repos="https://R-Forge.R-project.org")
}

## Conducting Installations from BioConductor

\dontrun{
source("https://bioconductor.org/biocLite.R")
}

## Reading the R Manual

}
}
\seealso{
\code{\link{toc}}

Other Chapters: 
\code{\link{ch01}()},
\code{\link{ch02}()},
\code{\link{ch03}()},
\code{\link{ch04}()},
\code{\link{ch05}()},
\code{\link{ch06}()},
\code{\link{ch07}()},
\code{\link{ch08}()},
\code{\link{ch09}()},
\code{\link{ch10}()},
\code{\link{ch11}()},
\code{\link{ch12}()},
\code{\link{ch13}()},
\code{\link{ch14}()},
\code{\link{ch15}()},
\code{\link{ch16}()},
\code{\link{ch17}()},
\code{\link{ch18}()},
\code{\link{ch19}()}
}
\concept{Chapters}
