% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ch20}
\alias{ch20}
\title{Print examples of chapter 20 of 'R for Dummies'.}
\usage{
ch20()
}
\description{
To print a listing of all examples of a chapter, use \code{ch20()}. To run all the examples of \code{ch20()}, use \code{example(ch20)}.
}
\examples{
# Chapter 20 - Ten Tips on Working with Packages

## Poking Around the Nooks and Crannies of CRAN

options("repos" = c(CRAN = "http://cran.ma.imperial.ac.uk/"))

## Finding Interesting Packages

## Installing Packages

## install.packages("fortunes")

## Loading Packages

library("fortunes")

## Reading the Package Manual and Vignette

library(help=fortunes)
vignette("fortunes")

## Updating Packages

## update.packages()

## Unloading Packages

search()
detach(package:fortunes, unload=TRUE)

## Forging Ahead with R-Forge

## install.packages("data.table", repos="http://R-Forge.R-project.org")

## Conducting Installations from BioConductor

source("http://bioconductor.org/biocLite.R")

## Reading the R Manual

}
\seealso{
toc

Other Chapters: \code{\link{ch01}}, \code{\link{ch1}};
  \code{\link{ch02}}, \code{\link{ch2}};
  \code{\link{ch03}}, \code{\link{ch3}};
  \code{\link{ch04}}, \code{\link{ch4}};
  \code{\link{ch05}}, \code{\link{ch5}};
  \code{\link{ch06}}, \code{\link{ch6}};
  \code{\link{ch07}}, \code{\link{ch7}};
  \code{\link{ch08}}, \code{\link{ch8}};
  \code{\link{ch09}}, \code{\link{ch9}};
  \code{\link{ch10}}; \code{\link{ch11}};
  \code{\link{ch12}}; \code{\link{ch13}};
  \code{\link{ch14}}; \code{\link{ch15}};
  \code{\link{ch16}}; \code{\link{ch17}};
  \code{\link{ch18}}; \code{\link{ch19}}
}

