% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapters-auto.R
\name{ch20}
\alias{ch20}
\title{Print examples of chapter 20 of 'R for Dummies'.}
\usage{
ch20()
}
\description{
To print a listing of all examples of a chapter, use \code{ch20()}. To run all the examples of \code{ch20()}, use \code{example(ch20)}.
}
\seealso{
\code{\link{toc}}

Other Chapters: \code{\link{ch01}}, \code{\link{ch02}},
  \code{\link{ch03}}, \code{\link{ch04}},
  \code{\link{ch05}}, \code{\link{ch06}},
  \code{\link{ch07}}, \code{\link{ch08}},
  \code{\link{ch09}}, \code{\link{ch10}},
  \code{\link{ch11}}, \code{\link{ch12}},
  \code{\link{ch13}}, \code{\link{ch14}},
  \code{\link{ch15}}, \code{\link{ch16}},
  \code{\link{ch17}}, \code{\link{ch18}},
  \code{\link{ch19}}
}

