% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{rfordummies}
\alias{rfordummies}
\alias{rfordummies-package}
\title{A package to accompany the book "R for Dummies".}
\description{
A package to accompany the book "R for Dummies".
}
\references{
de Vries, A. , & Meys, J. (2012). <i>R for dummies</i>. Chichester: Wiley. \url{http://rfordummies.com/}.
}

