% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfordummies-package.r
\docType{package}
\name{rfordummies}
\alias{rfordummies}
\title{A package to accompany the book "R for Dummies".}
\description{
This package contains all the code examples in the book 'R for Dummies' (2nd edition) by Andrie de Vries and Joris Meys.
}
\details{
To print the sample code for every chapter:
\itemize{
\item \code{\link[=ch1]{ch1()}}
\item \code{\link[=ch2]{ch2()}}
\item etc.
}

To print the table of contents:
\itemize{
\item \code{\link[=toc]{toc()}}
}

To save the \code{elements} data frame to either excel or csv format, use:
\itemize{
\item \code{\link[=saveElements]{saveElements()}}
}
}
\references{
de Vries, A. , & Meys, J. (2012). <i>R for dummies</i>. Chichester: Wiley. https://rfordummies.com/, ISBN-13: 978-1119962847.
}
